+++ src/azure-cli/azure/cli/command_modules/sql/_params.py
@@ -1433,10 +1433,16 @@ def load_arguments(self, _):
     with self.argument_context('sql midb restore') as c:
         create_args_for_complex_type(
             c, 'parameters', ManagedDatabase, [
+                'deleted_time',
                 'target_managed_database_name',
                 'target_managed_instance_name',
                 'restore_point_in_time'
             ])
+
+        c.argument('deleted_time',
+                   options_list=['--deleted-time'],
+                   help='If specified, restore from a deleted database instead of from an existing database.'
+                   ' Must match the deleted time of a deleted database on the source Managed Instance.')
 
         c.argument('target_managed_database_name',
                    options_list=['--dest-name'],
@@ -1464,8 +1470,29 @@ def load_arguments(self, _):
                    ' new database. Must be greater than or equal to the source database\'s'
                    ' earliestRestoreDate value. ' + time_format_help)
 
-    with self.argument_context('sql midb list') as c:
-        c.argument('managed_instance_name', id_part=None)
+    with self.argument_context('sql midb short-term-retention-policy set') as c:
+        create_args_for_complex_type(
+            c, 'parameters', ManagedDatabase, [
+                'deleted_time',
+                'retention_days'
+            ])
+
+        c.argument('deleted_time',
+                   options_list=['--deleted-time'],
+                   help='If specified, updates retention days for a deleted database, instead of an existing database.'
+                   'Must match the deleted time of a deleted database on the source Managed Instance.')
+
+        c.argument('retention_days',
+                   options_list=['--retention-days'],
+                   required=True,
+                   help='New backup short term retention policy in days.'
+                   'Valid policy for live database is 7-35 days, valid policy for dropped databases is 0-35 days.')
+
+    with self.argument_context('sql midb short-term-retention-policy show') as c:
+        c.argument('deleted_time',
+                   options_list=['--deleted-time'],
+                   help='If specified, shows retention days for a deleted database, instead of an existing database.'
+                   'Must match the deleted time of a deleted database on the source Managed Instance.')
 
     ###############################################
     #                sql virtual cluster          #